Regenerar código repositorio con nueva versión de pargo

+++ historias_de_usuario/sqliteust/s_historias_gen.go
@@ -3,7 +3,6 @@ package sqliteust
 import (
 	"database/sql"
 	"errors"
-	"strings"
 
 	"github.com/pargomx/gecko/gko"
 
@@ -11,68 +10,40 @@ import (
 )
 
 //  ================================================================  //
-//  ========== MYSQL/CONSTANTES ====================================  //
+//  ========== INSERT ==============================================  //
 
-// Lista de columnas separadas por coma para usar en consulta SELECT
-// en conjunto con scanRow o scanRows, ya que las columnas coinciden
-// con los campos escaneados.
-const columnasHistoria string = "historia_id, titulo, objetivo, prioridad, completada"
-
-// Origen de los datos de ust.Historia
-//
-// FROM historias
-const fromHistoria string = "FROM historias "
-
-//  ================================================================  //
-//  ========== MYSQL/TBL-INSERT ====================================  //
-
-// InsertHistoria valida el registro y lo inserta en la base de datos.
 func (s *Repositorio) InsertHistoria(his ust.Historia) error {
-	const op string = "mysqlust.InsertHistoria"
+	const op string = "InsertHistoria"
 	if his.HistoriaID == 0 {
-		return gko.ErrDatoInvalido().Msg("HistoriaID sin especificar").Ctx(op, "pk_indefinida")
+		return gko.ErrDatoIndef().Op(op).Msg("HistoriaID sin especificar").Str("pk_indefinida")
 	}
 	if his.Titulo == "" {
-		return gko.ErrDatoInvalido().Msg("Titulo sin especificar").Ctx(op, "required_sin_valor")
+		return gko.ErrDatoIndef().Op(op).Msg("Titulo sin especificar").Str("required_sin_valor")
 	}
-	err := his.Validar()
-	if err != nil {
-		return gko.ErrDatoInvalido().Err(err).Op(op).Msg(err.Error())
-	}
-	_, err = s.db.Exec("INSERT INTO historias "+
+	_, err := s.db.Exec("INSERT INTO historias "+
 		"(historia_id, titulo, objetivo, prioridad, completada) "+
 		"VALUES (?, ?, ?, ?, ?) ",
 		his.HistoriaID, his.Titulo, his.Objetivo, his.Prioridad, his.Completada,
 	)
 	if err != nil {
-		if strings.HasPrefix(err.Error(), "Error 1062 (23000)") {
-			return gko.ErrYaExiste().Err(err).Op(op)
-		} else if strings.HasPrefix(err.Error(), "Error 1452 (23000)") {
-			return gko.ErrDatoInvalido().Err(err).Op(op).Msg("No se puede insertar la información porque el registro asociado no existe")
-		} else {
-			return gko.ErrInesperado().Err(err).Op(op)
-		}
+		return gko.ErrAlEscribir().Err(err).Op(op)
 	}
 	return nil
 }
 
 //  ================================================================  //
-//  ========== MYSQL/TBL-UPDATE ====================================  //
+//  ========== UPDATE ==============================================  //
 
 // UpdateHistoria valida y sobreescribe el registro en la base de datos.
 func (s *Repositorio) UpdateHistoria(his ust.Historia) error {
-	const op string = "mysqlust.UpdateHistoria"
+	const op string = "UpdateHistoria"
 	if his.HistoriaID == 0 {
-		return gko.ErrDatoInvalido().Msg("HistoriaID sin especificar").Ctx(op, "pk_indefinida")
+		return gko.ErrDatoIndef().Op(op).Msg("HistoriaID sin especificar").Str("pk_indefinida")
 	}
 	if his.Titulo == "" {
-		return gko.ErrDatoInvalido().Msg("Titulo sin especificar").Ctx(op, "required_sin_valor")
+		return gko.ErrDatoIndef().Op(op).Msg("Titulo sin especificar").Str("required_sin_valor")
 	}
-	err := his.Validar()
-	if err != nil {
-		return gko.ErrDatoInvalido().Err(err).Op(op).Msg(err.Error())
-	}
-	_, err = s.db.Exec(
+	_, err := s.db.Exec(
 		"UPDATE historias SET "+
 			"historia_id=?, titulo=?, objetivo=?, prioridad=?, completada=? "+
 			"WHERE historia_id = ?",
@@ -86,14 +57,11 @@ func (s *Repositorio) UpdateHistoria(his ust.Historia) error {
 }
 
 //  ================================================================  //
-//  ========== MYSQL/TBL-DELETE ====================================  //
+//  ========== EXISTE ==============================================  //
 
-func (s *Repositorio) DeleteHistoria(HistoriaID int) error {
-	const op string = "mysqlust.DeleteHistoria"
-	if HistoriaID == 0 {
-		return gko.ErrDatoInvalido().Msg("HistoriaID sin especificar").Ctx(op, "pk_indefinida")
-	}
-	// Verificar que solo se borre un registro.
+// Retorna error nil si existe solo un registro con esta clave primaria.
+func (s *Repositorio) ExisteHistoria(HistoriaID int) error {
+	const op string = "ExisteHistoria"
 	var num int
 	err := s.db.QueryRow("SELECT COUNT(historia_id) FROM historias WHERE historia_id = ?",
 		HistoriaID,
@@ -105,53 +73,80 @@ func (s *Repositorio) DeleteHistoria(HistoriaID int) error {
 		return gko.ErrInesperado().Err(err).Op(op)
 	}
 	if num > 1 {
-		return gko.ErrInesperado().Err(nil).Op(op).Msgf("abortado porque serían borrados %v registros", num)
+		return gko.ErrInesperado().Err(nil).Op(op).Str("existen más de un registro para la pk").Ctx("registros", num)
 	} else if num == 0 {
-		return gko.ErrNoEncontrado().Err(ust.ErrHistoriaNotFound).Op(op).Msg("cero resultados")
+		return gko.ErrNoEncontrado().Err(ust.ErrHistoriaNotFound).Op(op)
 	}
-	// Eliminar registro
+	return nil
+}
+
+//  ================================================================  //
+//  ========== DELETE ==============================================  //
+
+func (s *Repositorio) DeleteHistoria(HistoriaID int) error {
+	const op string = "DeleteHistoria"
+	if HistoriaID == 0 {
+		return gko.ErrDatoIndef().Op(op).Msg("HistoriaID sin especificar").Str("pk_indefinida")
+	}
+	err := s.ExisteHistoria(HistoriaID)
+	if err != nil {
+		return gko.Err(err).Op(op)
+	}
 	_, err = s.db.Exec(
 		"DELETE FROM historias WHERE historia_id = ?",
 		HistoriaID,
 	)
 	if err != nil {
-		if strings.HasPrefix(err.Error(), "Error 1451 (23000)") {
-			return gko.ErrYaExiste().Err(err).Op(op).Msg("Este registro es referenciado por otros y no se puede eliminar")
-		} else {
-			return gko.ErrInesperado().Err(err).Op(op)
-		}
+		return gko.ErrAlEscribir().Err(err).Op(op)
 	}
 	return nil
 }
 
 //  ================================================================  //
-//  ========== MYSQL/SCAN-ROW ======================================  //
+//  ========== CONSTANTES ==========================================  //
+
+// Lista de columnas separadas por coma para usar en consulta SELECT
+// en conjunto con scanRow o scanRows, ya que las columnas coinciden
+// con los campos escaneados.
+//
+//	historia_id,
+//	titulo,
+//	objetivo,
+//	prioridad,
+//	completada
+const columnasHistoria string = "historia_id, titulo, objetivo, prioridad, completada"
+
+// Origen de los datos de ust.Historia
+//
+//	FROM historias
+const fromHistoria string = "FROM historias "
+
+//  ================================================================  //
+//  ========== SCAN ================================================  //
 
 // Utilizar luego de un sql.QueryRow(). No es necesario hacer row.Close()
-func (s *Repositorio) scanRowHistoria(row *sql.Row, his *ust.Historia, op string) error {
-
+func (s *Repositorio) scanRowHistoria(row *sql.Row, his *ust.Historia) error {
 	err := row.Scan(
 		&his.HistoriaID, &his.Titulo, &his.Objetivo, &his.Prioridad, &his.Completada,
 	)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			return gko.ErrNoEncontrado().Msg("Historia de usuario no se encuentra").Op(op)
+			return gko.ErrNoEncontrado().Msg("Historia de usuario no se encuentra")
 		}
-		return gko.ErrInesperado().Err(err).Op(op)
+		return gko.ErrInesperado().Err(err)
 	}
-
 	return nil
 }
 
 //  ================================================================  //
-//  ========== MYSQL/GET ===========================================  //
+//  ========== GET =================================================  //
 
 // GetHistoria devuelve un Historia de la DB por su clave primaria.
 // Error si no encuentra ninguno, o si encuentra más de uno.
 func (s *Repositorio) GetHistoria(HistoriaID int) (*ust.Historia, error) {
-	const op string = "mysqlust.GetHistoria"
+	const op string = "GetHistoria"
 	if HistoriaID == 0 {
-		return nil, gko.ErrDatoInvalido().Msg("HistoriaID sin especificar").Ctx(op, "pk_indefinida")
+		return nil, gko.ErrDatoIndef().Op(op).Msg("HistoriaID sin especificar").Str("pk_indefinida")
 	}
 	row := s.db.QueryRow(
 		"SELECT "+columnasHistoria+" "+fromHistoria+
@@ -159,11 +154,15 @@ func (s *Repositorio) GetHistoria(HistoriaID int) (*ust.Historia, error) {
 		HistoriaID,
 	)
 	his := &ust.Historia{}
-	return his, s.scanRowHistoria(row, his, op)
+	err := s.scanRowHistoria(row, his)
+	if err != nil {
+		return nil, err
+	}
+	return his, nil
 }
 
 //  ================================================================  //
-//  ========== MYSQL/SCAN-ROWS =====================================  //
+//  ========== SCAN ================================================  //
 
 // scanRowsHistoria escanea cada row en la struct Historia
 // y devuelve un slice con todos los items.
@@ -173,26 +172,40 @@ func (s *Repositorio) scanRowsHistoria(rows *sql.Rows, op string) ([]ust.Histori
 	items := []ust.Historia{}
 	for rows.Next() {
 		his := ust.Historia{}
-
 		err := rows.Scan(
 			&his.HistoriaID, &his.Titulo, &his.Objetivo, &his.Prioridad, &his.Completada,
 		)
 		if err != nil {
 			return nil, gko.ErrInesperado().Err(err).Op(op)
 		}
-
 		items = append(items, his)
 	}
 	return items, nil
 }
 
 //  ================================================================  //
-//  ========== MYSQL/LIST ==========================================  //
+//  ========== LIST ================================================  //
 
 func (s *Repositorio) ListHistorias() ([]ust.Historia, error) {
-	const op string = "mysqlust.ListHistorias"
+	const op string = "ListHistorias"
 	rows, err := s.db.Query(
 		"SELECT " + columnasHistoria + " " + fromHistoria,
+	)
+	if err != nil {
+		return nil, gko.ErrInesperado().Err(err).Op(op)
+	}
+	return s.scanRowsHistoria(rows, op)
+}
+
+//  ================================================================  //
+//  ========== LIST BYPADREID ======================================  //
+
+func (s *Repositorio) ListHistoriasByPadreID(nodoID int) ([]ust.Historia, error) {
+	const op string = "ListHistoriasByPadreID"
+	rows, err := s.db.Query(
+		"SELECT "+columnasHistoria+" "+fromHistoria+
+			"JOIN nodos ON nodo_id = historia_id WHERE padre_id = ? ORDER BY posicion",
+		nodoID,
 	)
 	if err != nil {
 		return nil, gko.ErrInesperado().Err(err).Op(op)
